Completed HW7

+++ DLEngine/src/DLEngine/Shaders/Hologram.hlsl
@@ -3,9 +3,9 @@
 
 struct InstanceInput
 {
-    float3 a_BaseColor     : BASE_COLOR;
-    float3 a_AddColor      : ADD_COLOR;
-    float a_TransformIndex : TRANSFORM_INDEX;
+    float3 a_BaseColor    : BASE_COLOR;
+    float3 a_AddColor     : ADD_COLOR;
+    uint a_TransformIndex : TRANSFORM_INDEX;
 };
 
 struct VertexOutput
@@ -21,7 +21,7 @@ VertexOutput mainVS(VertexInput vsInput, InstanceInput instInput)
 {
     VertexOutput vsOutput;
 
-    float4x4 meshToWorld = mul(c_MeshToModel, c_ModelToWorld[uint(instInput.a_TransformIndex)]);
+    float4x4 meshToWorld = mul(c_MeshToModel, c_ModelToWorld[instInput.a_TransformIndex]);
     float4 vertexPos = mul(float4(vsInput.a_Position, 1.0), meshToWorld);
 
     vsOutput.v_WorldPos = vertexPos.xyz;
